make payment_status_v2 private fnc
